requirements.txt
streamlit
pymongo
python-dotenv
psycopg2-binary
tqdm
pandas
tensorflow
tf_keras
keras
SQLAlchemy
huggingface-hub
transformers
deep_translator
langdetect
pdoc
scikit-learn
sentencepiece
sacremoses
nltk
pytest